+++ src/TwoBodyME.hh
@@ -166,7 +166,7 @@ class TwoBodyME
   void PrintAllMatrices() const;
   void PrintMatrix(size_t chbra,size_t chket) const;
 //  void PrintMatrix(size_t chbra,size_t chket) const {std::cout.precision(12); MatEl.at({chbra,chket}).raw_print();};
-  int Dimension();
+  size_t Dimension();
   size_t size();
 
   void WriteBinary(std::ofstream&);
